Update sync tests to use pytest [noissue]
