+++ pulp_deb/tests/functional/constants.py
@@ -61,6 +61,7 @@ DEB_FIXTURE_ARCH = "ppc64"
 DEB_FIXTURE_ARCH_UPDATE = "armeb"
 DEB_FIXTURE_STANDARD_REPOSITORY_NAME = "/debian/"
 DEB_FIXTURE_UPDATE_REPOSITORY_NAME = "/debian-update/"
+DEB_FIXTURE_INVALID_REPOSITORY_NAME = "/debian-invalid/"
 
 DEB_FIXTURE_SUMMARY = _clean_dict(
     {
